FAIR-3296: Move argument checking to executed script
Why:
* Possibly a case of over-refactoring in parent commit but the original code
  also did not handle the `code` argument well (still referring to it as the
  definition file in error message).
* Each script handles arguments differently so there's not much to be gained
  from making this part of the module
* Modules are more reusable outside of CLI context if we don't have specific
  argument formatting references and checks

Addresses need by:
* Moving argument checks to the directly executed scripts rather than modules

+++ fair-api-datasets-delete.py
@@ -1,15 +1,20 @@
 import requests
 import sys
-from common.constants import BASE_HEADERS, SSL_VERIFY
-from datasets.utilities import require_args, dataset_url
+from common.constants import BASE_HEADERS, DRY_RUN, SSL_VERIFY
+from datasets.utilities import dataset_url
 
-require_args()
+if len(sys.argv) < 2:
+    print(f'Usage: {sys.argv[0]} <code> <--dry-run>')
+    exit(1)
 
 dataset_code = sys.argv[1]
 url = dataset_url(dataset_code)
 
 print(f'API endpoint: {url}')
 print(f'Deleting Dataset with code: {dataset_code}')
+if DRY_RUN: 
+    print("DRY RUN - no requests sent")
+    exit(0)
 
 r = requests.delete(url, headers=BASE_HEADERS, verify=SSL_VERIFY)
 if r.status_code != 204:
